content-security-policy/script-src/script-src-strict_dynamic_meta_tag.html

<!DOCTYPE HTML>
<html>

<head>
    <title>A `strict-dynamic` policy can be served in a META tag.</title>
    <meta http-equiv="Content-Security-Policy" content="script-src 'strict-dynamic' 'nonce-dummy'">
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>
    <script src='../support/checkSecurityPolicyViolations.sub.js' nonce='dummy'></script>
    <!-- CSP served: script-src 'strict-dynamic' 'nonce-dummy' -->
</head>

<body>
    <h1>A `strict-dynamic` policy can be served in a META tag.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('script');
            e.src = 'simpleSourcedScript.js';
            e.onload = t.step_func_done(function() {
                assert_true(simpleSourcedScriptRan);
            });
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
        }, 'Script injected via `appendChild` is allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('script');
            e.src = 'simpleSourcedScript.js';
            e.setAttribute('nonce', 'wrong');
            e.onload = t.step_func_done(function() {
                assert_true(simpleSourcedScriptRan);
            });
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
        }, 'Script injected via `appendChild` is allowed with `strict-dynamic`, even if it carries an incorrect nonce.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('script');
            e.textContent = 'simpleSourcedScriptRan = true;';
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
            setTimeout(t.step_func_done(function() {
                assert_true(simpleSourcedScriptRan);
            }), 1000);
        }, 'Script injected via `appendChild` populated via `textContent` is allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('script');
            e.textContent = 'simpleSourcedScriptRan = true;';
            e.setAttribute('nonce', 'wrong');
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
            setTimeout(t.step_func_done(function() {
                assert_true(simpleSourcedScriptRan);
            }), 1000);
        }, 'Script injected via `appendChild` populated via `textContent` is allowed with `strict-dynamic`, even if it carries an incorrect nonce.');
    </script>

    <script nonce='dummy'>
        async_test(function(t) {
            setTimeout(t.step_func_done(function() {
                assert_no_violations();
            }), 1000);
        }, 'No CSP violation report has been fired.');
    </script>

</body>

</html>
